wip: testrpc restructuring
aesthetic refactor testrpc, cleanup

fixed most tests after rebase zkjs

fix encr test, update nullifier

ispublics
