+++ zk.js/src/rpc.ts
@@ -182,7 +182,8 @@ export class Rpc {
   }
 
   async getMerkleRoot(
-    _merkleTreePubkey: PublicKey,
+    /// For signature compatibility with TestRpc
+    _merkleTreePubkey?: PublicKey,
   ): Promise<{ root: string; index: number } | undefined> {
     throw new RpcError(
       RpcErrorCode.RPC_METHOD_NOT_IMPLEMENTED,
