+++ zk.js/src/transaction/index.ts
@@ -1,6 +1,6 @@
 export * from "./create-solana-transactions";
 export * from "./create-storage-transaction";
-export * from "./fetch-recent-transactions";
+export * from "./index-transactions";
 export * from "./psp-transaction";
 export * from "./send-versioned-transaction";
 export * from "./solana-transaction";
