+++ zk.js/src/utxo/index.ts
@@ -1,5 +1,6 @@
 export * from "./bn254";
 export * from "./create-out-utxos";
+export * from "./parse-utxo";
 export * from "./program-utxo";
 export * from "./select-in-utxos";
 export * from "./utxo";
